internal/config: add initial configuration file parsing
This leverages the `viper` library that allows configuration to be
parsed from a variety of sources. What I am using here is a TOML file
and environment variable overrides.

The first configuration parameter is the logging level for the `zap`
logger, which I plan to be using throughout this project.

This commit also adds the first Go code, and defines a Go module for
this project.