Makefile: add targets to run `gazelle` and `buildifier`
These are just for convenience. In one way, the `bazel-build` and
`bazel-test` targets are dependent on the `bazel-gazelle` target,
since the latter generates the build files that the former two depend
on. However, I want the developer to actively regenerate the build
files when needed (which will probably be helped by the build failing
if the build files are not regenerated). That way, the developer is
more conscious of when something in the build changes.

The purpose of the `bazel-buildifier` target is mostly to do linting
in CI -- if files are not formatted, the CI build should fail. The
`bazel-gazelle` target will also be used for that purpose.