Add `make` target to run the CI build locally
This will make it easier to develop and iterate the CI build
locally. The local CI environment has some limitations, though, such
as the inability to perform Docker commands. This is a bug/missing
feature of the `circleci` tool, and it would be difficult to work
around it for this specific repository.