Add `gofumports` as dependency
I chose to add this as a Go module dependency. Why do that, instead of
installing it with `gobin` (https://github.com/myitcv/gobin) or
similar? The reason is the following. While improperly formatted code
does not prevent the code to be _built_, it prevents the code from
being merged. As such, with the added benefit of not needing to
install another tool (`gobin`) to install yet another
tool (`gofumports`), I see it as a good tradeoff to include it in the
module dependencies.