Add `gobin` as an external tool
I eventually decided that running external tools such as code
formatters and linters (but not code generators such as `wire`) should
be done outside of Bazel. The reason was that I tried to add
`github.com/golangci/golangci-lint` as a tool dependency, but building
it with Bazel failed due to some transitive dependency importing an
external .proto file. In order to not have to deal with that
stuff (hopefully), I decided to run all external tools separately,
using `gobin`.