api/v1: add first version of a gRPC service, the User API
The User API will be used to manage the users of StreckU. Right now,
all it can do is to list users, but not create any users. As such, it
is not very valuable but it should be something to build upon.

With this commit, the generated Go code for all Protobuf messages and
services is also included. This is a conscious design decision -- to
version control generated code.

Pros:
* Checking out any commit in the version control history presents you
  with the complete view of the project at that point, including
  generated code. As such, the project is able to be built immediately
  after checking out. This makes things such as `git bisect` a lot
  easier, since you do not need to worry about regenerating code at
  each checkout.
* Committing generated code makes it clearer what code is actually run
  when the project is built and run.
Cons:
* Version control diffs can become huge as the generated code is
  updated. (1)
* There is now an extra step to have to remember when working on the
  project: if the source for code generation, e.g. Protobuf files, are
  changed, then code needs to be regenerated. It is possible to forget
  to regenerate code before committing other work, thereby leaving the
  project in an inconsistent state in version control. (2)

I think the pros outweigh the cons, and there are simple workaround
for the cons. For (1), one can just hide those diffs when reviewing a
pull request, viewing a commit in the commit log, etc. For (2), the CI
can be updated to regenerate code and check that the working directory
is clean afterwords -- such as what is already done with code
formatting.