Begin working on a `tools.mk` file to install all tools
My initial approach was to keep everything in a single, huge Makefile,
in order to make it easily searchable. However, this Makefile has
become difficult to overview and has introduced some problems relating
to dependencies between tools. As such, I have decided to move
everything out to a separate `tools.mk` Makefile that contains targets
and recipes for all tools. This will create clearer separation between
the installation of a tool, and the invocation of said tool. It will
also enable the main Makefile to have all tools be declared before
they are used, solving the problem of dependencies between tools.

In this initial iteration of `tools.mk`, the `gofumports` and
`golangci-lint` tools are the only tools included, just to show how
the `tools.mk` file should be structured.