tools.mk: specify `gobin` as order-only dependency
This will make caching in the CI environment work. The reason is that
for normal dependencies (i.e. not order-only dependencies), `make`
looks at the last modification time. When running in the CI
environment, `gobin` is downloaded each time, while tools from a
previous CI build might have been restored. Since `gobin` has a later
modification time than the tools, the tools are downloaded and
installed again, even though there is no need for it. With order-only
dependencies, last modification time does not matter, which means that
`make` will consider the tools to be up to date.