Add a simple integration test for the database connection pool
This commit adds an integration test that simply tries to create a new
database connection pool within a given timeout. This necessitates an
environment where a database is available, which is encoded in the
`integration_test.yml` file to be used with `docker-compose`. The
corresponding configuration values are added to a new
`configs/integration_test.toml` configuration file.