Begin work on an `internal/dockertest` package
The intention for this little package is to provide functionality to
run containers as part of integration testing. It is inspired by the
already-existing module `github.com/ory/dockertest`, which I have
tried but felt was a bit too bulky for my use case. This is the
beginning of an experiment to see whether I can create a package that
will fulfill this project's needs for integration testing without
jumping through too many hoops.