internal/dockertest: add methods to start and stop a container
The `StartContainer` method will also pull the image (which will
shortcircuit if the image already exists), as well as create the
container before starting it. The `StopContainer` method
will also remove the container.

I decided to include this functionality in these methods, while not
exposing it in the names, since the context where this package will be
used in is the context of throwaway containers. As such, one can
expect the container to be created and removed after one is done with
them.