internal/dockertest: use the `hello-world` image in tests
This should hopefully make these tests a little less heavy-weight to
run in a clean environment, such as CI. While the `alpine` flavors of
the Postgres image are typically small, the `hello-world` image is
much smaller still. However, most likely, the `alpine` flavors of the
Postgres image will be used in actual integration tests where this
package will be used for launching a Postgres container.