internal/dockertest: remove `WithContainer` method
I figured that if you want to use multiple containers in a test, then
the `WithContainer` pattern would give way to right-drifting
code. Since we can view a container as a resource that has to be
closed, just like a reader or some other IO resource, we might as well
handle it in the same way we usually handle readers: by closing them
manually in a `defer`red func or something.