internal/dockertest: move all methods over to `Client`
I figured: what is the need for having a separate "pool" structure?
What I'm providing an abstraction over/convenience methods for is the
Docker client.

Furthermore, the underlying Docker client is no longer exported. This
is to avoid confusion as to why the wrapper `Client` both has a
`ContainerStart` and `StartContainer` method, for example.