backend: add in-memory implementation of `RoleAPI`
I am intentionally fairly slacking with request validation since this
is only a toy example so far. For example, I managed to cause the
implementation to perform a nil dereference by not specifying a `Role`
message in a call to `CreateRole`. This will be improved in the
future, either by changing the API to have `required` fields, or by
programming more defensively.