api,backend: add a `ResetAPI` for testing
The purpose of this API is explained in the comment for the `ResetAPI`
service. The intention is to use it in integration testing, allowing
for the state to be reset between tests. This should decouple the
integration tests from the implementation, so that the tests and the
implementation can be written in different languages. Furthermore, it
will more accurately test the public API of the services.