backend: use build tags for integration testing
In this way, one has to explicitly opt in to integration testing
functionality at compile-time, rather than at runtime. I think this
solution is a little more elegant, and it also removes the risk of
accidentally running the backend in an environment where the
integration testing functionality definitely _not_ should be enabled.