testing: add a first set of tests for `UserAPI`
I am not entirely content with these tests. For example, the test for
`GetUser` requires that `ListUsers` works correctly. However, I see no
other way to do this without changing the APIs to accept a "desired"
ID, or to somehow fix a random seed. Therefore, this will have to do
for now.