api: set Go package to proper import path
This was previously not done as I had not figured out how to generate
code in a module-aware mode. Without the `module` option given to the
code generator plugins, the code would have been generated at an
import path of something like
`github.com/Saser/strecku/github.com/saser/strecku/v1`, which of
course is not correct.