internal/repositories: define Users repository and write test suite for it
The users interface is defined since I intend to have two
implementations: one in-memory implementation (like the
users.Repository type) and at least one database implementation. The
in-memory implementation can be used when one just wants to write
tests for something that uses the Users repository or when a fast,
throwaway repository (perhaps helpful when iterating on the UI).

The test suite is more or less a port of the tests for
users.Repository but adapted to the new Users interface.